[Feat/#22] 영수증 사진 받아 처리하는 로직 임시 구현

+++ Refrigerator/src/main/java/com/example/project/domain/expense/dto/ExpenseRequest.java
@@ -9,7 +9,7 @@ import jakarta.validation.constraints.Min;
 public record ExpenseRequest(
 	    LocalDateTime spentAt,
 	    @Min(value = 0, message = "금액은 0원 이상이어야 합니다.")
-	    Integer amount,
+	    Long amount,
 	    String title,
 	    Category category,
 	    String memo
